requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dateutil
pyinstaller
pywebview